sstreamを修正

+++ core/common/sstream.cpp
@@ -15,8 +15,8 @@ StringStream::StringStream()
 
 void  StringStream::checkSize(size_t size)
 {
-    if (size > (int)m_buffer.size())
-        while ((int)m_buffer.size() < size)
+    if (size > m_buffer.size())
+        while (m_buffer.size() < size)
             m_buffer.push_back(0);
 }
 
@@ -25,10 +25,10 @@ int  StringStream::read(void *buf, int count)
     if (m_pos == m_buffer.size())
         throw StreamException("End of stream");
 
-    int bytesRead = (std::min)((size_t)count, m_buffer.size() - m_pos);
+    size_t bytesRead = (std::min)(static_cast<size_t>(count), m_buffer.size() - m_pos);
     memcpy(buf, m_buffer.c_str() + m_pos, bytesRead);;
     m_pos += bytesRead;
-    return bytesRead;
+    return static_cast<int>(bytesRead);
 }
 
 void StringStream::write(const void *buf, int count)
@@ -61,12 +61,12 @@ void StringStream::seekTo(int newPos)
 
 int  StringStream::getPosition()
 {
-    return (int)m_pos;
+    return static_cast<int>(m_pos);
 }
 
 int  StringStream::getLength()
 {
-    return (int)m_buffer.size();
+    return static_cast<int>(m_buffer.size());
 }
 
 std::string StringStream::str()
